src/tile_processor.hpp: having strange problems when not using the 64-bit values for accumulating quality sums

+++ src/tile_processor.hpp
@@ -62,7 +62,7 @@ struct tile_processor {
   };
 
   // ADS: needs to count roughly ~1M reads each contributing up to 128
-  using qual_vec = std::vector<std::pair<std::uint32_t, std::uint32_t>>;
+  using qual_vec = std::vector<std::pair<std::uint64_t, std::uint64_t>>;
   static constexpr auto read_skip = 10 - 1;
 
   static std::uint32_t preceding_colons;
